Remove text objects for bullets
It was cool as a learning tool to understand how vim works but really
did not care to use those - so they were not maintained and updated, nor
do I think anybody else was, and they were conflicting with built-in vim
mappings. So many reasons why I'm happy to see those gone.

[Closes #45]