Make bullet list item styles configurable
This commit introduces a new option `g:bullets_list_item_styles` to
specify a list of regexes that match valid bullet point characters to
be recognized by bullets.vim.

The default is the same as the currently hardcoded list.

This commit only introduces this configurability for bullet list items
as this may be the most useful bullet item type to have configurability.

It may be considered to provide this configurability for the other bullet
item types, too (like checkbox list items, etc.).